Refactor proccessError and extract extractError
